+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Consolidate `import-export-comma-style` and `diff-friendly-import-export` into a new option `import-export-style` ([#201](https://github.com/fourmolu/fourmolu/pull/207))
+
 ## Fourmolu 0.7.0.1
 
 * Fix bad copy/paste where parsing errors for `haddock-style` would mention `CommaStyle`